fix: rename resolverBuilder to resolverMapper
